Omaha #3392 Change Doubles to Float in ACARS and ACARSSoundingLayer

Former-commit-id: 7d352f77a3822098dc18b94b00d03ff30771517d [formerly 06b9d253f354fd6af4da4333386cb406077bfd97] [formerly 7d352f77a3822098dc18b94b00d03ff30771517d [formerly 06b9d253f354fd6af4da4333386cb406077bfd97] [formerly 4f3a729894c53c13fad3b5b7e994c328cff6496f [formerly 9ac070996c42b004f3683d7ee253feb54f05264c]]]
Former-commit-id: 4f3a729894c53c13fad3b5b7e994c328cff6496f
Former-commit-id: 0067bc04d317142c7a44c36947ee724ff4e3ba48 [formerly 085fe0e47917dcc7ee3c3a9496c29c344174f0d4]
Former-commit-id: 0bfeb9fd7f327df47400cf6e94cb64d89b4277fb

+++ edexOsgi/com.raytheon.uf.edex.plugin.acars/src/com/raytheon/uf/edex/plugin/acars/decoder/ACARSDataAdapter.java
@@ -1,19 +1,19 @@
 /**
  * This software was developed and / or modified by Raytheon Company,
  * pursuant to Contract DG133W-05-CQ-1067 with the US Government.
- * 
+ *
  * U.S. EXPORT CONTROLLED TECHNICAL DATA
  * This software product contains export-restricted data whose
  * export/transfer/disclosure is restricted by U.S. law. Dissemination
  * to non-U.S. persons whether in the United States or abroad requires
  * an export license or other authorization.
- * 
+ *
  * Contractor Name:        Raytheon Company
  * Contractor Address:     6825 Pine Street, Suite 340
  *                         Mail Stop B8
  *                         Omaha, NE 68106
  *                         402.291.0100
- * 
+ *
  * See the AWIPS II Master Rights File ("Master Rights File.pdf") for
  * further licensing information.
  **/
@@ -46,9 +46,9 @@ import com.raytheon.uf.edex.decodertools.bufr.packets.IBUFRDataPacket;
 
 /**
  * Adapter used to decode ACARS data in BUFR format.
- * 
+ *
  * <pre>
- * 
+ *
  * SOFTWARE HISTORY
  * Date          Ticket#  Engineer    Description
  * ------------- -------- ----------- --------------------------
@@ -58,9 +58,10 @@ import com.raytheon.uf.edex.decodertools.bufr.packets.IBUFRDataPacket;
  * Mar 27, 2014  2811     skorolev    Added check for empty message.
  * May 14, 2014  2536     bclement    moved WMO Header to common, removed TimeTools usage
  * Jun 12, 2014  2061     bsteffen    Generate unique stationid
- * 
+ * Jul 22, 2014  3392     nabowle     ACARSRecord has Float fields instead of Double
+ *
  * </pre>
- * 
+ *
  * @author jkorman
  * @version 1.0
  */
@@ -90,7 +91,7 @@ public class ACARSDataAdapter {
     private String traceId = null;
 
     /**
-     * 
+     *
      * @param name
      */
     @Deprecated
@@ -98,7 +99,7 @@ public class ACARSDataAdapter {
     }
 
     /**
-     * 
+     *
      */
     public ACARSDataAdapter() {
     }
@@ -155,7 +156,7 @@ public class ACARSDataAdapter {
     }
 
     /**
-     * 
+     *
      * @param data
      * @return
      */
@@ -452,7 +453,7 @@ public class ACARSDataAdapter {
     }
 
     /**
-     * 
+     *
      * @param packets
      * @param locPos
      * @return
@@ -506,7 +507,7 @@ public class ACARSDataAdapter {
     }
 
     /**
-     * 
+     *
      * @param packets
      * @param locPos
      * @return
@@ -554,7 +555,7 @@ public class ACARSDataAdapter {
     }
 
     /**
-     * 
+     *
      * @param dataList
      * @param record
      * @return
@@ -566,7 +567,7 @@ public class ACARSDataAdapter {
         int d = packet.getReferencingDescriptor().getDescriptor();
         if (d == BUFRDescriptor.createDescriptor(0, 12, 1)) {
             if (!packet.isMissing()) {
-                record.setTemp((Double) packet.getValue());
+                record.setTemp(((Double) packet.getValue()).floatValue());
             }
         }
 
@@ -582,7 +583,7 @@ public class ACARSDataAdapter {
         d = packet.getReferencingDescriptor().getDescriptor();
         if (d == BUFRDescriptor.createDescriptor(0, 11, 2)) {
             if (!packet.isMissing()) {
-                record.setWindSpeed((Double) packet.getValue());
+                record.setWindSpeed(((Double) packet.getValue()).floatValue());
             }
         }
 
@@ -697,7 +698,8 @@ public class ACARSDataAdapter {
                 d = packet.getReferencingDescriptor().getDescriptor();
                 if (d == BUFRDescriptor.createDescriptor(0, 11, 2)) {
                     if (!packet.isMissing()) {
-                        record.setWindSpeed((Double) packet.getValue());
+                        record.setWindSpeed(((Double) packet.getValue())
+                                .floatValue());
                     }
                 }
 
@@ -714,7 +716,8 @@ public class ACARSDataAdapter {
                 d = packet.getReferencingDescriptor().getDescriptor();
                 if (d == BUFRDescriptor.createDescriptor(0, 12, 101)) {
                     if (!packet.isMissing()) {
-                        record.setTemp((Double) packet.getValue());
+                        record.setTemp(((Double) packet.getValue())
+                                .floatValue());
                     }
                 }
 
@@ -722,7 +725,8 @@ public class ACARSDataAdapter {
                 d = packet.getReferencingDescriptor().getDescriptor();
                 if (d == BUFRDescriptor.createDescriptor(0, 12, 103)) {
                     if (!packet.isMissing()) {
-                        record.setDwpt((Double) packet.getValue());
+                        record.setDwpt(((Double) packet.getValue())
+                                .floatValue());
                     }
                 }
             }
@@ -751,7 +755,7 @@ public class ACARSDataAdapter {
         d = packet.getReferencingDescriptor().getDescriptor();
         if (d == BUFRDescriptor.createDescriptor(0, 11, 2)) {
             if (!packet.isMissing()) {
-                record.setWindSpeed((Double) packet.getValue());
+                record.setWindSpeed(((Double) packet.getValue()).floatValue());
             }
         }
 
@@ -759,7 +763,7 @@ public class ACARSDataAdapter {
         d = packet.getReferencingDescriptor().getDescriptor();
         if (d == BUFRDescriptor.createDescriptor(0, 12, 1)) {
             if (!packet.isMissing()) {
-                record.setTemp((Double) packet.getValue());
+                record.setTemp(((Double) packet.getValue()).floatValue());
             }
         }
 
@@ -767,7 +771,7 @@ public class ACARSDataAdapter {
         d = packet.getReferencingDescriptor().getDescriptor();
         if (d == BUFRDescriptor.createDescriptor(0, 13, 2)) {
             if (!packet.isMissing()) {
-                record.setMixingRatio((Double) packet.getValue());
+                record.setMixingRatio(((Double) packet.getValue()).floatValue());
             }
         }
 
@@ -775,7 +779,7 @@ public class ACARSDataAdapter {
         d = packet.getReferencingDescriptor().getDescriptor();
         if (d == BUFRDescriptor.createDescriptor(0, 13, 3)) {
             if (!packet.isMissing()) {
-                record.setHumidity((Double) packet.getValue());
+                record.setHumidity(((Double) packet.getValue()).floatValue());
             }
         }
 
@@ -824,7 +828,7 @@ public class ACARSDataAdapter {
         d = packet.getReferencingDescriptor().getDescriptor();
         if (d == BUFRDescriptor.createDescriptor(0, 11, 2)) {
             if (!packet.isMissing()) {
-                record.setWindSpeed((Double) packet.getValue());
+                record.setWindSpeed(((Double) packet.getValue()).floatValue());
             }
         }
 
@@ -849,7 +853,7 @@ public class ACARSDataAdapter {
         d = packet.getReferencingDescriptor().getDescriptor();
         if (d == BUFRDescriptor.createDescriptor(0, 12, 101)) {
             if (!packet.isMissing()) {
-                record.setTemp((Double) packet.getValue());
+                record.setTemp(((Double) packet.getValue()).floatValue());
             }
         }
 
@@ -898,8 +902,8 @@ public class ACARSDataAdapter {
     }
 
     /**
-     * 
-     * 
+     *
+     *
      * @param dataList
      * @param pos
      * @return
@@ -973,7 +977,7 @@ public class ACARSDataAdapter {
         int d = packet.getReferencingDescriptor().getDescriptor();
         if (d == BUFRDescriptor.createDescriptor(0, 7, 4)) {
             if (!packet.isMissing()) {
-                record.setPressure((Double) packet.getValue());
+                record.setPressure(((Double) packet.getValue()).floatValue());
             }
         }
     }
